Add support for Apple MPS GPU
This adds support for Apple MPS GPUs (Apple Silicon).  It adds a new function called to_gpu_mps, modelled on the to_gpu function.  A default parameter (__init__) is set to False.

It is based on the guidance below from Apple (see the 'verify' section):
https://developer.apple.com/metal/pytorch/

I have not been able to test this because it connects onto Google Collab and other files I do not have.